Fix unreliability issue
The problem was that the chip-select CS pin was not configured, so it was
floating (!).

Obviously this caused problems for operation such as extreme noise
sensitivity etc.

With this fix, it seems to make sense to go back to 25MHz IO drive
strength. There might be a cycle or two that could be shaved from
timings, but spec sets a maximum of 15 MHz operation, and there seems
little need to go marginal on the timings, things are plenty fast like
this.